共有された旅程表のURL構造を変更
