+++ app/(memorybook)/memorybook/(share)/share/[trip_id]/layout.tsx
